requirements.txt
gradio
transformers
torch
pillow
ffmpeg-python
